- Removed legacy size handling:   - Deleted SizeSeeder and all size-related migration files (sizes, product_size).   - Removed all foreign key constraints and tables related to size. - Updated attribute system:   - Ensured 'attribute_values' table supports all flexible value columns (text_value, boolean_value, etc.).   - Updated AttributeFactory and AttributeSystemSeeder to use 'is_filterable' and 'is_configurable' fields.   - Switched seeder to use firstOrCreate for attributes to avoid unique constraint errors. - All product attribute logic is now handled through the flexible attribute system.

+++ Modules/Core/Helpers/Helper.php
@@ -29,7 +29,7 @@ class Helper
     {
         $user_id = self::getUserId($user_id);
 
-        return $user_id !== 0 ? Cart::whereUserId($user_id)->whereOrderId(null)->sum('quantity') : 0;
+        return $user_id !== 0 ? (int) Cart::whereUserId($user_id)->whereOrderId(null)->sum('quantity') : 0;
     }
 
     public static function getAllProductFromWishlist(string $user_id = ''): Collection
